split functions in different files
split add,minus,mult,divide,expo in their respective files.
