+++ src/functions/mod.rs
@@ -1,0 +1,6 @@
+pub mod add;
+pub mod divide;
+pub mod expo;
+pub mod function;
+pub mod minus;
+pub mod mult;
