+++ src/interpreting/mod.rs
@@ -1,3 +1,2 @@
-pub mod function;
 pub(crate) mod interpreter;
 pub mod stdlib;
